refactor(webapp): drop onNotification prop; widgets call useNotification() inline
onNotification was a pure pass-through of the useNotification() wrapper from the Panes; the wrapper drives debounce, cross-promo toasts, linked-action suppression, and rate-card injection — all webapp-layer logic widgets can call directly now that they're webapp-internal. Twelve widgets read the callback inline. WidgetMessage is promoted to export and a new OnNotificationCallback alias is added; eighteen transaction-callback hooks switch from `Pick<WidgetProps, 'onNotification'>` to an inline `onNotification?: OnNotificationCallback` field. WagmiWrapper gains a MemoryRouter wrap so usePrepareNotification's react-router reads work in tests. A latent circular-import (`@/widgets` re-entering itself mid-load) was fixed by reordering the barrel constants-first and switching useNotification/useWidgetAnalytics/useUserSuggestedActions to deep import paths. PsmConversionWidget's bare-render test gains an inline mock rather than pulling in the new router/wagmi stack. Typecheck/lint/build green; 222 tests pass.
