+++ apps/webapp/src/widgets/StUSDSWidget/hooks/useStUsdsTransactionCallbacks.tsx
@@ -4,7 +4,7 @@ import { t } from '@lingui/core/macro';
 import { formatUnits } from 'viem';
 import { useTransactionCallbacks } from '@/widgets/shared/hooks/useTransactionCallbacks';
 import { TransactionCallbacks } from '@/widgets/shared/types/transactionCallbacks';
-import { WidgetProps } from '@/widgets/shared/types/widgetState';
+import { WidgetProps, OnNotificationCallback } from '@/widgets/shared/types/widgetState';
 import { WidgetAnalyticsEvent, WidgetAnalyticsEventType } from '@/widgets/shared/types/analyticsEvents';
 import { useMemo, useRef } from 'react';
 import { useChainId } from 'wagmi';
@@ -12,8 +12,9 @@ import { StUSDSAction, StUSDSFlow } from '../lib/constants';
 
 interface UseStUsdsTransactionCallbacksParameters extends Pick<
   WidgetProps,
-  'addRecentTransaction' | 'onWidgetStateChange' | 'onNotification' | 'onAnalyticsEvent'
+  'addRecentTransaction' | 'onWidgetStateChange' | 'onAnalyticsEvent'
 > {
+  onNotification?: OnNotificationCallback;
   amount: bigint;
   needsAllowance: boolean;
   shouldUseBatch: boolean;
