+++ apps/webapp/src/widgets/UpgradeWidget/hooks/useUpgradeTransactionCallbacks.tsx
@@ -4,15 +4,16 @@ import { formatUnits } from 'viem';
 import { useChainId } from 'wagmi';
 import { useTransactionCallbacks } from '@/widgets/shared/hooks/useTransactionCallbacks';
 import { TransactionCallbacks } from '@/widgets/shared/types/transactionCallbacks';
-import { WidgetProps } from '@/widgets/shared/types/widgetState';
+import { WidgetProps, OnNotificationCallback } from '@/widgets/shared/types/widgetState';
 import { WidgetAnalyticsEvent, WidgetAnalyticsEventType } from '@/widgets/shared/types/analyticsEvents';
 import { useMemo, useRef } from 'react';
 import { UpgradeAction, UpgradeFlow } from '../lib/constants';
 
 interface UseUpgradeTransactionCallbacksParameters extends Pick<
   WidgetProps,
-  'addRecentTransaction' | 'onWidgetStateChange' | 'onNotification' | 'onAnalyticsEvent'
+  'addRecentTransaction' | 'onWidgetStateChange' | 'onAnalyticsEvent'
 > {
+  onNotification?: OnNotificationCallback;
   originAmount: bigint;
   originToken: Token;
   targetToken: Token;
